Add Marshaler and Unmarshaler interfaces
Add Marshaler and Unmarshaler interfaces, which allow for custom encoding/decoding
working similarly to encoding/json Marshaler and Unmarshaler:

```
type Marshaler interface {
	MarshalXDR(e *Encoder) (int, error)
}

type Unmarshaler interface {
	UnmarshalXDR(e *Decoder) (int, error)
}
```

The Decoder and Encoder automatically call the relevant methods if found.

Both implementation by value and by pointer are supported.

A new error (`ErrCustomUnaddressableByPointer`) will be returned if
a by-pointer implementation is found but the value passed is unaddressable
(not allowing to call the method by pointer).